requirements.txt
decorator==4.0.6
Django==1.9.1
djangorestframework==3.3.2
ipdb==0.8.1
ipython==4.0.1
ipython-genutils==0.1.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
simplegeneric==0.8.1
traitlets==4.0.0
